update(lucifer-front): move project details in an other component
