+++ apps/lucifer-front/src/projects/ProjectPage.tsx
@@ -1,44 +1,62 @@
-import React, { FC, useCallback, useEffect, useState } from 'react';
-import { useHistory, useParams } from 'react-router';
-import { useForm } from 'react-hook-form';
-import { IUpdateProject } from '@lucifer/types';
+import React, { FC, useCallback, useRef, useState } from 'react';
+import { useHistory, useParams, useRouteMatch } from 'react-router';
 
-import { CircularProgress, Fab, Grid, IconButton, Paper, TextField, Typography, Zoom } from '@material-ui/core';
+import { IconButton, Paper, Tab, Tabs } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import { Delete as DeleteIcon, Save as SaveIcon } from '@material-ui/icons';
-import { LabelledText, RefreshButton } from '@lucifer/react/basics';
+import { Delete as DeleteIcon } from '@material-ui/icons';
+import { RefreshButton } from '@lucifer/react/basics';
 
 import { useNeedRole } from '../auth/auth.hooks';
 
 import { useProject } from './projects.hooks';
 import { ProjectHeader } from './ProjectHeader';
+import { Link as RouterLink } from 'react-router-dom';
+import { ProjectDetailsTab } from './ProjectDetailsTab';
 
-// Styles
-const useStyles = makeStyles(({ spacing }) => ({
-  root: {
-    padding: spacing(3),
-  },
-  description: {
-    whiteSpace: 'pre-line'
-  },
-  save: {
-    position: 'absolute',
-    bottom: spacing(2),
-    right: spacing(2)
-  }
-}));
+// Utils
+interface LinkTabProps {
+  value: string;
+  label: string;
+  disabled?: boolean;
+}
+
+const LinkTab = (props: LinkTabProps) => {
+  const { value } = props;
+  const { url } = useRouteMatch();
+  const { page } = useParams<ProjectParams>();
+
+  return (
+    <Tab {...props}
+      component={RouterLink}
+      to={page ? url.replace(page, value) : `${url}/${value}`}
+    />
+  );
+};
 
 // Types
 interface ProjectParams {
   userId: string;
   id: string;
+  page: string;
 }
+
+// Styles
+const useStyles = makeStyles({
+  toolbar: {
+    display: 'grid',
+    justifyItems: 'end',
+
+    '& > *': {
+      gridArea: '1 / 1 / 2 / 2',
+    }
+  }
+});
 
 // Component
 export const ProjectPage: FC = () => {
   // Router
   const history = useHistory();
-  const { userId, id } = useParams<ProjectParams>();
+  const { userId, id, page = "details" } = useParams<ProjectParams>();
 
   // API
   const { project, loading, reload, update, remove } = useProject(userId, id);
@@ -46,23 +64,8 @@ export const ProjectPage: FC = () => {
   // Auth
   const isAdmin = useNeedRole('admin', usr => project?.adminId === usr?.id) ?? false;
 
-  // Form
-  const { errors, register, reset, handleSubmit, formState } = useForm<IUpdateProject>({
-    mode: 'onChange'
-  });
-
   // State
   const [isRemoving, setRemoving] = useState(false);
-
-  // Effects
-  useEffect(() => {
-    if (project) {
-      reset({
-        name: project.name,
-        description: project.description
-      });
-    }
-  }, [reset, project]);
 
   // Callbacks
   const handleDelete = useCallback(async () => {
@@ -70,6 +73,9 @@ export const ProjectPage: FC = () => {
     await remove();
     history.goBack();
   }, [remove, history]);
+
+  // Refs
+  const actionsContainer = useRef<HTMLDivElement>(null);
 
   // Render
   const styles = useStyles();
@@ -80,66 +86,24 @@ export const ProjectPage: FC = () => {
         <ProjectHeader
           project={project}
           actions={(
-            <>
-              <IconButton disabled={!isAdmin || isRemoving} onClick={handleDelete}>
-                <DeleteIcon />
-              </IconButton>
-              <RefreshButton disabled={isRemoving} refreshing={loading} onClick={reload} />
-            </>
+            <div className={styles.toolbar} ref={actionsContainer}>
+              <span>
+                <IconButton disabled={!isAdmin || isRemoving} onClick={handleDelete}>
+                  <DeleteIcon />
+                </IconButton>
+                <RefreshButton disabled={isRemoving} refreshing={loading} onClick={reload} />
+              </span>
+            </div>
           )}
         />
+        <Tabs variant="fullWidth" value={page} onChange={() => null}>
+          <LinkTab value="details" label="Détails" />
+        </Tabs>
       </Paper>
-      <form className={styles.root} onSubmit={handleSubmit(update)}>
-        { project && (
-          <Grid container spacing={4} direction="column">
-            <Grid item container spacing={2}>
-              <Grid item xs={12} sm={6} md={4}>
-                { isAdmin ? (
-                  <TextField
-                    label="Nom" variant="outlined" fullWidth
-                    name="name" inputRef={register({
-                      maxLength: { value: 100, message: '100 charactères max.' },
-                    })}
-                    error={!!errors.name} helperText={errors.name?.message}
-                  />
-                ) : (
-                  <LabelledText label="Nom">
-                    <Typography>{ project.name }</Typography>
-                  </LabelledText>
-                ) }
-              </Grid>
-            </Grid>
-            <Grid item container spacing={2}>
-              <Grid item xs>
-                { isAdmin ? (
-                  <TextField
-                    label="Description" variant="outlined" fullWidth multiline
-                    name="description" inputRef={register}
-                    error={!!errors.description} helperText={errors.description?.message}
-                  />
-                ) : (
-                  <LabelledText label="Description">
-                    <Typography className={styles.description}>{ project.description }</Typography>
-                  </LabelledText>
-                ) }
-              </Grid>
-            </Grid>
-          </Grid>
-        ) }
-        { isAdmin && (
-          <Zoom in appear>
-            <Fab
-              className={styles.save} color="primary"
-              type="submit" disabled={!formState.isDirty || formState.isSubmitting || isRemoving}
-            >
-              <SaveIcon />
-            </Fab>
-          </Zoom>
-        ) }
-        { formState.isSubmitting && (
-          <CircularProgress className={styles.save} size={56} />
-        ) }
-      </form>
+      <ProjectDetailsTab
+        project={project} show={page === "details"} isRemoving={isRemoving}
+        onUpdate={update}
+      />
     </>
   );
 };
